num of element

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -6,7 +6,7 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t num_of_elements = 0;
+	int num_of_elements = 0;
 
 	while (h != NULL)
 	{
